requirements.txt
flask==3.0.3
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
gunicorn==22.0.0
httpx==0.27.0
pydantic==2.8.2
python-dotenv==1.0.1
cirpy==1.0.2
requests>=2.32.0
pydantic-settings>=2.3.4
numpy==1.26.4
rdkit-pypi==2022.9.5
pandas==2.2.2
openpyxl==3.1.5
xlrd>=2.0.1
smbprotocol==1.15.0
matplotlib==3.9.0
pymupdf>=1.24.0
Pillow>=10.0.0
portkey-ai>=1.0.0
